+++ src/collective/cover/testing.py
@@ -83,6 +83,12 @@ class Fixture(PloneSandboxLayer):
         #self.loadZCML(package=plone.app.imagetile)
         #import plone.app.texttile
         #self.loadZCML(package=plone.app.imagetile)
+        if 'virtual_hosting' not in app.objectIds():
+            # If ZopeLite was imported, we have no default virtual
+            # host monster
+            from Products.SiteAccess.VirtualHostMonster \
+                import manage_addVirtualHostMonster
+            manage_addVirtualHostMonster(app, 'virtual_hosting')
 
     def setUpPloneSite(self, portal):
         # Install into Plone site using portal_setup