+++ src/collective/cover/tests/test_acceptance.py
@@ -3,31 +3,13 @@
 from collective.cover.testing import FUNCTIONAL_TESTING
 from plone.testing import layered
 
+import os
 import robotsuite
 import unittest
 
-import pkg_resources
-PLONE_VERSION = pkg_resources.require("Plone")[0].version
-
-tests = [
-    'test_basic_tile.txt',
-    'test_collection_tile.txt',
-    'test_contentchooser_search_tab.txt',
-    'test_contenttree_tab.txt',
-    'test_contenttree_tab_path.txt',
-    'test_cover.txt',
-    'test_embed_tile.txt',
-    'test_file_tile.txt',
-    'test_image_tile.txt',
-    'test_layout.txt',
-    'test_list_tile.txt',
-    'test_locked_cover.txt',
-]
-
-# FIXME: test randomly failing under Plone 4.3.x
-#        see https://github.com/collective/collective.cover/issues/155
-if '4.3' in PLONE_VERSION:
-    tests.remove('test_contentchooser_search_tab.txt',)
+dirname = os.path.dirname(__file__)
+files = os.listdir(dirname)
+tests = [f for f in files if f.startswith('test_') and f.endswith('.txt')]
 
 
 def test_suite():