+++ src/collective/cover/tests/test_embed_tile.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
 
-import unittest2 as unittest
-
+from collective.cover.testing import INTEGRATION_TESTING
+from collective.cover.tiles.base import IPersistentCoverTile
+from collective.cover.tiles.embed import EmbedTile
 from zope.interface.verify import verifyClass
 from zope.interface.verify import verifyObject
 
-from collective.cover.testing import INTEGRATION_TESTING
-from collective.cover.tiles.embed import EmbedTile
-from collective.cover.tiles.base import IPersistentCoverTile
+import unittest
 
 
 class EmbedTileTestCase(unittest.TestCase):
@@ -19,6 +18,8 @@ class EmbedTileTestCase(unittest.TestCase):
         self.request = self.layer['request']
         self.cover = self.portal['frontpage']
         self.tile = EmbedTile(self.cover, self.request)
+        # XXX: tile initialization
+        self.tile.__name__ = 'collective.cover.embed'
 
     def test_interface(self):
         self.assertTrue(IPersistentCoverTile.implementedBy(EmbedTile))