+++ src/collective/cover/tests/test_image_tile.py
@@ -46,7 +46,17 @@ class ImageTileTestCase(unittest.TestCase):
         obj = self.portal['my-image']
         self.tile.populate_with_object(obj)
         rendered = self.tile()
-        self.assertTrue('test-image-tile/@@images' in rendered)
+        self.assertIn('test-image-tile/@@images', rendered)
+
+    @unittest.expectedFailure
+    def test_alt_atribute_present_in_image(self):
+        """Object's title must be displayed in image alt attribute.
+        See: https://github.com/collective/collective.cover/issues/182
+        """
+        obj = self.portal['my-image']
+        self.tile.populate_with_object(obj)
+        rendered = self.tile()
+        self.assertIn('alt="Test image"', rendered)
 
     def test_image_traverser(self):
         obj = self.portal['my-image']