+++ src/collective/cover/tests/test_link_tile.py
@@ -54,3 +54,13 @@ class LinkTileTestCase(unittest.TestCase):
         self.assertTrue('Test link' in rendered)
         self.assertTrue(
             "This link was created for testing purposes" in rendered)
+
+    @unittest.expectedFailure
+    def test_alt_atribute_present_in_image(self):
+        """Object's title must be displayed in image alt attribute.
+        See: https://github.com/collective/collective.cover/issues/182
+        """
+        obj = self.portal['my-link']
+        self.tile.populate_with_object(obj)
+        rendered = self.tile()
+        self.assertIn('alt="Test link"', rendered)